feat: Implement VRF subscription management and deployment flow
- Add FundSubscription and AddConsumer contracts for VRF setup
- Update DeployRaffle to handle subscription creation, funding, and consumer registration
- Add Link token support in HelperConfig
- Update subscriptionId type from uint64 to uint256 for compatibility
- Add broadcast directory to .gitignore
- Update foundry.toml with solmate remapping and fs permissions
- Update foundry-devops dependency
- Fix VRF coordinator address for Sepolia network

+++ test/unit/RaffleTest.t.sol
@@ -16,7 +16,7 @@ contract RaffleTest is Test {
     uint256 public interval;
     address public vrfCoordinator;
     bytes32 public gasLane;
-    uint64 public subscriptionId;
+    uint256 public subscriptionId;
     uint256 public callbackGasLimit;
 
     /* Events */
